Alter app's logic & structure
