--- src/containers/about/index.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-const About = () => (
-  <div>
-    <h1>About Page</h1>
-    <p>Did you get here via Redux?</p>
-  </div>
-);
-
-export default About;
