+++ src/index.js
@@ -2,8 +2,8 @@ import React from "react";
 import { render } from "react-dom";
 import { Provider } from "react-redux";
 import { ConnectedRouter } from "react-router-redux";
-import store, { history } from "./store";
-import App from "./containers/app";
+import store, { history } from "./store/store";
+import App from "./containers/app/App";
 
 import "sanitize.css/sanitize.css";
 import "./index.css";
